pyproject.toml
[tool.poetry]
name = "hypergraph-scattering"
version = "0.1.0"
description = "Introduction of the hypergraph scattering transform, with options for various wavelets and variations of the scattering transform"
authors = ["HungryAmoeba <charles.xu@yale.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
numpy = "^1.26.1"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.0"
argparse = "^1.4.0"
networkx = "^3.2.1"
phate = "^1.0.11"
pyyaml = "^6.0.1"
dhg = "^0.9.3"
einops = "^0.7.0"
torch-geometric = "^2.4.0"
ipykernel = "^6.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
